Email MicroService Refactoring
- Encapsulate Email config in a separate file
- Makes EmailMicroservice independent
