+++ emailMicroService/index.js
@@ -1,13 +1,13 @@
 'use strict';
-const cote = require('cote');
+
+const { Requester } = require('cote');
 const sgMail = require('@sendgrid/mail');
 require('dotenv').config();
 
 sgMail.setApiKey(process.env.SENDGRID_API_KEY);
-
 console.log('SENDGRID_API_KEY:', process.env.SENDGRID_API_KEY);
 
-const emailResponder = new cote.Responder({ name: 'email-responder' });
+const emailResponder = new Responder({ name: 'email-responder' });
 
 emailResponder.on('send_email', async (req, done) => {
   const {
