+++ emailMicroService/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "emailmicroservice",
+  "version": "1.0.0",
+  "description": "Welcome email Microservice",
+  "main": "index.js",
+  "scripts": {
+    "start": "node index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@sendgrid/mail": "^7.7.0",
+    "cote": "^1.0.3",
+    "dotenv": "^16.3.1"
+  }
+}
