+++ models/User.js
@@ -1,10 +1,5 @@
 const mongoose = require('mongoose');
 const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
-const cote = require('cote');
-const sgMail = require('@sendgrid/mail');
-
-const requester = new cote.Requester({ name: 'email-requester' });
 
 //DONE Create Schema Users
 
@@ -45,51 +40,17 @@ userSchema.statics.findUserById = function (userId) {
   return query;
 };
 
-//DONE method that sends an email that contains the user's own email and a token to change the password
-userSchema.statics.microEmailService = async function (to) {
-  try {
-    const user = await User.findOne({ email: to });
-
-    if (!user) {
-      throw new Error('user not found');
-    }
-
-    //NOTE We clear the resetpassword field before storing the new value
-    user.resetpassword = '';
-
-    //NOTE Here you generate the password recovery token
-    const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET, {
-      expiresIn: '1h',
-    });
-
-    //NOTE I delete the old token before re-saving the new one
-    user.resetpassword = token;
-    await user.save();
-
-    //NOTE Send the message to the microservice to send the email
-    const emailRequest = {
-      type: 'send_email',
-      to: to,
-      from: process.env.EMAIL_SEND_GRID,
-      templateId: process.env.TEMPLATE_ID_SEND_GRID,
-      dynamic_template_data: {
-        email: to,
-        token: token,
-      },
-    };
-
-    return new Promise(resolve => requester.send(emailRequest, resolve));
-  } catch (error) {
-    throw new Error('Error sending password recovery email.');
-  }
-};
-
-//DONE Static method to reset password
-
+/**
+ * This reset user password
+ *
+ * @param {*} email
+ * @param {*} token
+ * @param {*} newPassword
+ * @returns message of password was succesfully reset or throws error if it fails
+ */
 userSchema.statics.resetPassword = async function (email, token, newPassword) {
   try {
-    //NOTE Search for the user in the database by their email
-
+    // Search for the user in the database by their email
     const user = await this.findOne({ email: email, resetpassword: token });
 
     //NOTE If the user is not found or the token does not match, return an error
